Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1939 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/widget_manager.cpp
@@ -829,7 +829,7 @@ int WidgetManager::calcLineX( const int POS )
     int width = 0;
     WidgetID* curr_wgt = 0;
 
-    const int NUM_ELEMS = m_elems.size();
+    const int NUM_ELEMS = (int)m_elems.size();
 
     for( int i = POS; i < NUM_ELEMS; ++i )
     {
